alterando regra do campo de ownerId da tabela grupos para deixar o campo como opicional.

+++ src/users/users.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateUserDTO } from './dto/create-user.dto';
 import { UpdateUserDTO } from './dto/update-user.dto';
@@ -80,11 +84,24 @@ export class UsersService {
   }
 
   async remove(id: string) {
-    // Verifica se o usuário existe antes de deletar
-    const user = await this.prisma.user.findUnique({ where: { id } });
+    const user = await this.prisma.user.findUnique({
+      where: { id },
+      include: { ownedGroups: true }, // Busca os grupos que ele possui
+    });
+
     if (!user) {
       throw new NotFoundException(`Usuário com ID ${id} não encontrado`);
     }
+
+    if (user.ownedGroups.length > 0) {
+      throw new BadRequestException(
+        `Usuário ${user.name} não pode ser excluído porque ainda é responsavel por grupos`,
+      );
+    }
+
+    await this.prisma.userGroup.deleteMany({
+      where: { userId: id },
+    });
 
     return this.prisma.user.delete({
       where: { id },
